refactor(graph): extract ConversationManager to js/graph/conversation-manager.js and import in app.js
