+++ app.js
@@ -953,186 +953,8 @@ function stableLinkId(aId, bId) {
   return `link_${x}_${y}`;
 }
 
-// Conversation manager: link multiple copilots to share history and enforce turn-taking
-const ConversationManager = (() => {
-  const conversations = new Map(); // convId -> { members:Set<number>, history:[], turnIdx:number, pending:[], busy:boolean }
-
-  function ensureConv(convId) {
-    if (conversations.has(convId)) return conversations.get(convId);
-    const conv = { members: new Set(), history: [], turnIdx: 0, pending: [], busy: false };
-    conversations.set(convId, conv);
-    return conv;
-  }
-  function create() {
-    const id = 'conv_' + Date.now() + '_' + Math.random().toString(36).slice(2, 6);
-    ensureConv(id);
-    return id;
-  }
-  function addMember(convId, copilot) {
-    const conv = ensureConv(convId);
-    conv.members.add(copilot.id);
-    copilot._convId = convId;
-  }
-  function link(a, b) {
-    const convId = a._convId || b._convId || create();
-    addMember(convId, a);
-    addMember(convId, b);
-    return convId;
-  }
-  function getMembers(convId) {
-    const conv = conversations.get(convId);
-    return conv ? Array.from(conv.members) : [];
-  }
-  function getHistory(convId) {
-    const conv = conversations.get(convId);
-    return conv ? conv.history : null;
-  }
-  function recordAssistant(convId, assistantMsg) {
-    const conv = conversations.get(convId);
-    if (!conv) return;
-    conv.history.push({ role: 'assistant', content: assistantMsg });
-  }
-
-  function renderUserInMembers(convId, msg, senderId) {
-    const ids = getMembers(convId);
-  const authorName = getGlobalUserName();
-    ids.forEach(id => {
-      const inst = CopilotManager.instances.get(id);
-      if (!inst) return;
-      if (id === senderId) {
-        // sender usually already rendered; ensure at least once
-        // no-op
-      } else {
-        inst.addUser(msg, authorName);
-      }
-    });
-  }
-  function renderAssistantInMembers(convId, msg, authorName) {
-    const ids = getMembers(convId);
-    ids.forEach(id => {
-      const inst = CopilotManager.instances.get(id);
-      if (!inst) return;
-      // In raw mode, prefer typing effect; in md, render formatted immediately
-      if (typeof inst.renderAssistantReply === 'function') {
-        inst.renderAssistantReply(msg, authorName);
-      } else if ((inst.renderMode || 'raw') === 'md') {
-        inst.addAssistant(msg, authorName);
-      } else {
-        inst.addAssistant(msg, authorName);
-      }
-    });
-  }
-
-  async function process(convId) {
-    const conv = conversations.get(convId);
-    if (!conv || conv.busy) return;
-    if (PauseManager.isPaused()) return; // globally paused
-    if (!conv.pending.length) return;
-    conv.busy = true;
-    // round-robin select next responder
-    const ids = getMembers(convId);
-    if (!ids.length) { conv.busy = false; return; }
-    conv.turnIdx = (conv.turnIdx + 1) % ids.length;
-  const nextId = ids[conv.turnIdx];
-  const responder = CopilotManager.instances.get(nextId) || CopilotManager.instances.values().next().value;
-    // highlight
-    CopilotManager.instances.forEach(inst => inst.panel.classList.remove('active-speaking'));
-    if (responder) responder.panel.classList.add('active-speaking');
-    // Use full shared history as context
-    const messages = conv.history.slice();
-    try {
-  const reply = await responder.generateReply(messages);
-      const text = reply || '(inget svar)';
-  recordAssistant(convId, text);
-  renderAssistantInMembers(convId, text, responder ? responder.name : undefined);
-    } catch (e) {
-  renderAssistantInMembers(convId, 'Nätverksfel.', responder ? responder.name : undefined);
-    } finally {
-      if (responder) responder.panel.classList.remove('active-speaking');
-      conv.busy = false;
-      if (conv.pending.length) {
-        // slight delay to keep UI smooth
-        setTimeout(() => process(convId), 20);
-      }
-    }
-  }
-
-  function enqueueUser(sender, msg) {
-    const conv = conversations.get(sender._convId);
-    if (!conv) return;
-    if (PauseManager.isPaused()) {
-      // still record user and render, but don't start processing
-      conv.history.push({ role: 'user', content: msg });
-      renderUserInMembers(sender._convId, msg, sender.id);
-      conv.pending.push({ from: sender.id, msg });
-      toast('Flöde pausat – meddelandet köades.', 'warn');
-      return;
-    }
-    // record user message in shared history and render in all members
-    conv.history.push({ role: 'user', content: msg });
-    renderUserInMembers(sender._convId, msg, sender.id);
-    // queue a turn
-    conv.pending.push({ from: sender.id, msg });
-    if (!conv.busy) process(sender._convId);
-  }
-
-  function resumeAll() {
-    // Attempt to process any conversations with pending items
-    conversations.forEach((conv, id) => {
-      if (conv.pending.length && !conv.busy) process(id);
-    });
-  }
-
-  function removeMember(convId, copilot) {
-    const conv = conversations.get(convId);
-    if (!conv) return;
-    conv.members.delete(copilot.id);
-    // If conversation now has fewer than 2 members, dissolve it and
-    // merge its remaining shared history back into the single member (if any),
-    // then remove the conversation to avoid self-replies.
-    const size = conv.members.size;
-    if (size <= 0) {
-      conversations.delete(convId);
-      return;
-    }
-    if (size === 1) {
-      // Get the remaining member
-      const remainingId = Array.from(conv.members)[0];
-      const remainingInst = CopilotManager.instances.get(remainingId);
-      if (remainingInst) {
-        // Append shared history to the remaining copilot's personal history
-        try {
-          for (const m of conv.history || []) {
-            if (m && m.role) remainingInst.history.push(m);
-          }
-        } catch (e) {}
-        // Clear any pending turns and prevent further processing
-        conv.pending = [];
-        conv.busy = false;
-  // Clear the remaining instance's conv pointer so it no longer references a deleted conv
-  try { remainingInst._convId = null; } catch (e) {}
-      }
-      conversations.delete(convId);
-      return;
-    }
-    // More than one member: keep conversation and adjust turnIdx
-    conv.turnIdx = conv.turnIdx % size;
-  }
-
-  function removePendingFor(copilotId) {
-    conversations.forEach((conv, id) => {
-      if (!conv || !Array.isArray(conv.pending)) return;
-      const before = conv.pending.length;
-      conv.pending = conv.pending.filter(p => p && p.from !== copilotId);
-      if (conv.pending.length !== before) {
-        // if queue changed and there are no pending items left, ensure not busy
-        if (!conv.pending.length) conv.busy = false;
-      }
-    });
-  }
-
-  return { link, addMember, getMembers, getHistory, recordAssistant, enqueueUser, removeMember, resumeAll, removePendingFor };
-})();
+// Conversation manager moved to module
+import { ConversationManager } from './js/graph/conversation-manager.js';
 
 class CopilotInstance {
   constructor(id, opts = {}) {
